additional abstract base
